Version 2.3.0

+++ Src/OpenEditor/OEStorage.cpp
@@ -526,7 +526,7 @@ void Storage::GetLine  (int line, const char*& ptr) const
 
 void Storage::GetLine (int line, const char*& ptr, int& len) const
 {
-  if(line >= 0 && line < m_Lines.size())
+  if(line >= 0 && line < (int)m_Lines.size())
   {
     ptr = m_Lines.at(line).data();
     len = (int)m_Lines[line].length();
